chore(embedder): align cubert pipeline under service and update import
